requirements.txt
Django==1.8.2
dj-database-url==0.3.0
django-suit==0.2.13
python-decouple==2.3
